refactor(card): extract the image decode/reveal gate into useImageReveal
Move card-cover's decode-then-reveal logic (skeleton until decoded, then fade in,
re-run on source change / element reinsert) into a colocated composable so it can
be shared and card-cover stays declarative.
